(PC_36196): reduce cognitivecomplexity
